Finalizando Backend

+++ backend/src/controllers/IncidentController.js
@@ -3,7 +3,24 @@ const connection = require('../database/connection');
 
 module.exports = {
   async index(request, response) {
-    const incidents = await connection('incidents').select('*');
+    const { page = 1 } = request.query;
+
+    const [count] = await connection('incidents').count();
+
+    const incidents = await connection('incidents')
+      .join('ongs', 'ong_id', '=', 'incidents.ong_id')
+      .limit(5)
+      .offset((page - 1) * 5)
+      .select([
+        'incidents.*',
+        'ongs.name',
+        'ongs.email',
+        'ongs.whatsapp',
+        'ongs.city',
+        'ongs.uf'
+      ]);
+
+    response.header('X-Total-Count', count['count(*)']);
   
     return response.json(incidents);
   },
@@ -28,10 +45,12 @@ module.exports = {
 
     const incident = await connection('incidents').where('id', id).select('ong_id').first();
 
+    if(!incident) {
+      return response.status(400).json({ error: 'Incident not found.' });
+    }
+
     if(incident && incident.ong_id !== ong_id) {
       return response.status(401).json({ error: 'Operation not permitted.' });
-    } else {
-      return response.status(400).json({ error: 'Incident not found.' });
     }
 
     await connection('incidents').where('id', id).delete();
